+++ packages/parser/.eslintignore
@@ -1,4 +1,4 @@
 node_modules
 lib
 doc
-src/ast/grammar/dot.pegjs.ts
+src/dot.pegjs.ts
