--- packages/parser/src/ast/index.ts
@@ -1,2 +1,0 @@
-export * from './parse';
-export * from './types';
