+++ packages/parser/src/index.ts
@@ -1,3 +1,3 @@
-export * from './parser';
-export * from './usecase';
-export * as ast from './ast';
+export { SyntaxError } from './dot.pegjs';
+export { AST } from './ast';
+export { parse, dot } from './usecase';
